roles

+++ app/Http/Controllers/Admin/ServiceController.php
@@ -17,7 +17,11 @@ class ServiceController extends Controller
      */
     public function index()
     {
-        $data = Service::latest()->paginate(10);
+        if (Auth::user()->can('all-services')) {
+            $data = Service::latest()->paginate(10);
+        }elseif(Auth::user()->can('services')){
+            $data = Service::where('admin_id',Auth::user()->id)->latest()->paginate(10);   
+        }
         return view('admin.services.index',compact('data'));
     }
 
@@ -54,7 +58,11 @@ class ServiceController extends Controller
 
     public function show(string $id)
     {
-        $service = Service::findOrFail($id);
+        if (Auth::user()->can('all-services')) {
+            $service = Service::findOrFail($id);
+        }elseif(Auth::user()->can('services')){
+            $service = Service::where('admin_id',Auth::user()->id)->findOrFail($id);;   
+        }
         return view('admin.services.show',compact('service'));
     }
 
@@ -63,7 +71,11 @@ class ServiceController extends Controller
      */
     public function edit(string $id)
     {
-        $service = Service::findOrFail($id);
+        if (Auth::user()->can('all-services')) {
+            $service = Service::findOrFail($id);
+        }elseif(Auth::user()->can('services')){
+            $service = Service::where('admin_id',Auth::user()->id)->findOrFail($id);;   
+        }
         $categories = Category::sub()->get();
         return view('admin.services.edit',compact('service','categories'));
     }
@@ -101,7 +113,11 @@ class ServiceController extends Controller
      */
     public function destroy(Request $request)
     {
-        Service::findOrFail($request->id)->delete();
+        if (Auth::user()->can('all-services')) {
+
+            Service::findOrFail($request->id)->delete();
+        }
+        
         return redirect()->route('admin.services.index')->with('success','Service has been removed successfully');
     }
 }
